feat: Implement webapp API server with SQLite support and Docker Compose setup
- Added main application entry point for the webapp API server.
- Implemented graceful shutdown and database migration for SQLite.
- Created Docker Compose configuration for development environment including API, SQLite, and frontend services.
- Updated documentation to reflect new Docker Compose file location and usage instructions.
- Enhanced logging and error handling throughout the application.
- Added structured logging and observability features.
- Refactored repository and service layers for better organization and clarity.
- Updated HTTP transport layer to include new routes and middleware.
- Improved API key authentication and session management.
- Added health check endpoints for the API server.
- Introduced new environment variables for configuration.
- Updated frontend to align with backend changes, including API request structure.
