+++ .golangci.yml
@@ -5,15 +5,44 @@ run:
   tests: true
 
 linters:
-  default: none
-  enable:
-  - misspell
-  - govet
-  - staticcheck
-  - errcheck
-  - ineffassign
-  - unused
-  - revive
+  default: all
+  disable:
+    - wsl
+    - wsl_v5
+    - nlreturn
+    - varnamelen
+    - wrapcheck
+    - funcorder
+    - godot
+    - exhaustruct
+    - embeddedstructfieldcheck
+    - depguard
+    - dupl
+    - lll
+    - mnd
+    - noinlineerr
+    - testpackage
+    - err113
+    - errchkjson
+    - funlen
+    - goconst
+    - gocritic
+    - intrange
+    - nestif
+    - noctx
+    - paralleltest
+    - unconvert
+  settings:
+    govet:
+      enable:
+        - nilness
+        - printf
+        - lostcancel
+        - sortslice
+    tagliatelle:
+      case:
+        rules:
+          json: snake
 
 formatters:
   enable:
