--- docker-compose.yml
@@ -1,108 +1,0 @@
-version: '3.8'
-
-services:
-  # Main API service - Development mode with live reloading
-  api:
-    image: golang:1.22-alpine
-    container_name: webapp-api-dev
-    platform: linux/arm64
-    working_dir: /app
-    volumes:
-      - ./:/app
-      - webapp_cache:/go/pkg/mod  # Go module cache
-      - dbdata:/db
-    environment:
-      - HTTP_ADDRESS=0.0.0.0:8082
-      - DB_DRIVER=sqlite
-      - DB_PATH=/db/app.db
-      - CORS_ALLOWED_ORIGINS=http://web:5173,http://localhost:5173,http://instance-agent.subnet05071228.vcn05071228.oraclevcn.com:5173
-      - ENV=dev
-      - LOG_LEVEL=debug
-      - LOG_FORMAT=text
-      - GOTOOLCHAIN=go1.22.6
-    command: sh -c "apk add --no-cache git wget && go mod download && go run ./cmd/webapp"
-    depends_on:
-      db:
-        condition: service_started
-    expose:
-      - "8082"
-    networks:
-      - webapp-network
-    healthcheck:
-      test: ["CMD-SHELL", "wget -qO- http://127.0.0.1:8082/healthz >/dev/null 2>&1 || exit 1"]
-      interval: 10s
-      timeout: 5s
-      retries: 10
-      start_period: 60s
-
-  # SQLite database service for development
-  db:
-    image: arm64v8/alpine:3.19
-    container_name: webapp-db-dev
-    platform: linux/arm64
-    command: sh -c "apk add --no-cache sqlite sqlite-libs >/dev/null 2>&1 && sqlite3 /data/app.db 'PRAGMA user_version;' || true; tail -f /dev/null"
-    volumes:
-      - dbdata:/data
-    networks:
-      - webapp-network
-    healthcheck:
-      test: ["CMD-SHELL", "sqlite3 /data/app.db 'PRAGMA user_version;' >/dev/null 2>&1 || sqlite3 /data/app.db '.databases' >/dev/null 2>&1"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-      start_period: 10s
-
-  # Frontend React/TypeScript application
-  web:
-    image: node:18-alpine
-    container_name: webapp-web
-    platform: linux/arm64
-    working_dir: /app/web
-    volumes:
-      - ./web:/app/web
-      - webapp_node_modules:/app/web/node_modules
-    environment:
-      - NODE_ENV=development
-      - VITE_PROXY_TARGET=http://api:8082
-    command: sh -c "npm ci && npm run dev -- --host 0.0.0.0 --port 5173"
-    ports:
-      - "5173:5173"
-    networks:
-      - webapp-network
-    depends_on:
-      api:
-        condition: service_healthy
-
-  # SQLite Admin - Development tool (profile: tools)
-  sqlite-admin:
-    image: python:3.12-alpine
-    container_name: webapp-sqlite-admin
-    platform: linux/arm64
-    working_dir: /app
-    volumes:
-      - dbdata:/data
-    command: sh -c "pip install --no-cache-dir sqlite-web && sqlite_web --host 0.0.0.0 --port 8080 /data/app.db"
-    ports:
-      - "8080"
-    networks:
-      - webapp-network
-    profiles:
-      - tools
-    depends_on:
-      db:
-        condition: service_started
-    healthcheck:
-      test: ["CMD-SHELL", "wget -qO- http://127.0.0.1:8080 >/dev/null 2>&1 || exit 1"]
-      interval: 10s
-      timeout: 5s
-      retries: 10
-      start_period: 20s
-
-volumes:
-  dbdata:
-  webapp_cache:
-  webapp_node_modules:
-
-networks:
-  webapp-network:
-    driver: bridge
