+++ internal/service/service.go
@@ -1,19 +1,23 @@
+// Package service implements business logic orchestrating repositories.
 package service
 
 import (
 	"context"
 	"errors"
 
-	"github.com/personal/webapp/internal/domain"
-	"github.com/personal/webapp/internal/repository"
+	"github.com/mdco1990/webapp/internal/domain"
+	"github.com/mdco1990/webapp/internal/repository"
 )
 
+// Service exposes business operations.
 type Service struct {
 	repo *repository.Repository
 }
 
+// New creates a Service backed by the provided repository.
 func New(repo *repository.Repository) *Service { return &Service{repo: repo} }
 
+// ErrValidation is returned when inputs fail validation.
 var ErrValidation = errors.New("validation error")
 
 func validateYM(ym domain.YearMonth) error {
@@ -23,6 +27,7 @@ func validateYM(ym domain.YearMonth) error {
 	return nil
 }
 
+// SetSalary validates inputs then stores salary for a month.
 func (s *Service) SetSalary(ctx context.Context, ym domain.YearMonth, amount domain.Money) error {
 	if err := validateYM(ym); err != nil {
 		return err
@@ -33,6 +38,7 @@ func (s *Service) SetSalary(ctx context.Context, ym domain.YearMonth, amount dom
 	return s.repo.UpsertSalary(ctx, ym, amount)
 }
 
+// SetBudget validates inputs then stores budget for a month.
 func (s *Service) SetBudget(ctx context.Context, ym domain.YearMonth, amount domain.Money) error {
 	if err := validateYM(ym); err != nil {
 		return err
@@ -43,6 +49,7 @@ func (s *Service) SetBudget(ctx context.Context, ym domain.YearMonth, amount dom
 	return s.repo.UpsertBudget(ctx, ym, amount)
 }
 
+// AddExpense validates and creates an expense.
 func (s *Service) AddExpense(ctx context.Context, e *domain.Expense) (int64, error) {
 	if err := validateYM(domain.YearMonth{Year: e.Year, Month: e.Month}); err != nil {
 		return 0, err
@@ -53,6 +60,7 @@ func (s *Service) AddExpense(ctx context.Context, e *domain.Expense) (int64, err
 	return s.repo.AddExpense(ctx, e)
 }
 
+// ListExpenses returns expenses for a month.
 func (s *Service) ListExpenses(ctx context.Context, ym domain.YearMonth) ([]domain.Expense, error) {
 	if err := validateYM(ym); err != nil {
 		return nil, err
@@ -60,6 +68,7 @@ func (s *Service) ListExpenses(ctx context.Context, ym domain.YearMonth) ([]doma
 	return s.repo.ListExpenses(ctx, ym)
 }
 
+// DeleteExpense removes an expense by ID.
 func (s *Service) DeleteExpense(ctx context.Context, id int64) error {
 	if id <= 0 {
 		return ErrValidation
@@ -67,6 +76,7 @@ func (s *Service) DeleteExpense(ctx context.Context, id int64) error {
 	return s.repo.DeleteExpense(ctx, id)
 }
 
+// Summary returns aggregate info for a month.
 func (s *Service) Summary(ctx context.Context, ym domain.YearMonth) (domain.Summary, error) {
 	if err := validateYM(ym); err != nil {
 		return domain.Summary{}, err
