+++ sonar-project.properties
@@ -1,0 +1,14 @@
+sonar.projectKey=mdco1990_WebApp
+sonar.organization=mdco1990
+
+
+# This is the name and version displayed in the SonarCloud UI.
+#sonar.projectName=WebApp
+#sonar.projectVersion=1.0
+
+
+# Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
+#sonar.sources=.
+
+# Encoding of the source code. Default is default system encoding
+#sonar.sourceEncoding=UTF-8
